requirements.txt
asgiref==3.6.0
decouple==0.0.7
Django==4.1.8
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
drf-flex-fields==1.0.2
environ==1.0
psycopg2==2.9.6
pymongo==4.3.3
pytz==2023.3
sqlparse==0.2.4
